[QUZ-90]{FEATURE] Quiz 옵션별 선택 수 구현

+++ quiz-service/quiz-domain/src/main/kotlin/com/grepp/quizy/quiz/domain/quiz/Quiz.kt
@@ -33,7 +33,7 @@ sealed class Quiz(
         }
     }
 
-    fun getTotalAnsweredCount(): Int {
+    fun getTotalAnsweredCount(): Long {
         return content.options.sumOf { it.selectionCount }
     }
